requirements.txt
apipkg==1.4
attrs==16.0.0
bintrees==2.0.4
croniter==0.3.12
execnet==1.4.1
freezegun==0.3.7
py==1.4.31
pytest==3.0.1
pytest-cache==1.0
python-dateutil==2.5.3
six==1.10.0
